Refs 9309. Move ChangeForReader to its own header file.
