+++ include/fastdds/rtps/writer/ReaderProxy.h
@@ -20,22 +20,26 @@
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS_PUBLIC
 
-#include <algorithm>
-#include <mutex>
-#include <set>
-#include <atomic>
+#include <fastdds/rtps/attributes/WriterAttributes.h>
+#include <fastdds/rtps/attributes/RTPSParticipantAllocationAttributes.hpp>
 
 #include <fastdds/rtps/builtin/data/ReaderProxyData.h>
-#include <fastdds/rtps/writer/ReaderLocator.h>
 
 #include <fastdds/rtps/common/Types.h>
 #include <fastdds/rtps/common/Locator.h>
 #include <fastdds/rtps/common/SequenceNumber.h>
 #include <fastdds/rtps/common/CacheChange.h>
 #include <fastdds/rtps/common/FragmentNumber.h>
-#include <fastdds/rtps/attributes/WriterAttributes.h>
-#include <fastdds/rtps/attributes/RTPSParticipantAllocationAttributes.hpp>
+
+#include <fastdds/rtps/writer/ChangeForReader.h>
+#include <fastdds/rtps/writer/ReaderLocator.h>
+
 #include <fastrtps/utils/collections/ResourceLimitedVector.hpp>
+
+#include <algorithm>
+#include <mutex>
+#include <set>
+#include <atomic>
 
 namespace eprosima {
 namespace fastrtps {
